# StudentPortal.API/bruno/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StudentPortal API",
    "description": "PRN232 Lab 1 – RESTful API for managing Students, Courses, Subjects and Semesters.",
    "contact": {
      "name": "PRN232 Student",
      "email": "student@fpt.edu.vn"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/enrollments": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Get all enrollments with pagination, filtering, $select and $expand.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortdescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$expand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a paged list of enrollments.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectPagedResponseApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Enroll a student into a course.",
        "requestBody": {
          "description": "StudentId and CourseId.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Enrollment created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentBaseResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, student/course not found, or already enrolled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/enrollments/{id}": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Get a single enrollment by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Enrollment ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrollment found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentBaseResponseApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Enrollment not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Delete an enrollment by ID (unenroll a student).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Enrollment ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrollment deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Enrollment not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEnrollmentRequest": {
        "required": [
          "courseId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "studentId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "courseId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 1
          }
        },
        "additionalProperties": false
      },
...
      "EnrollmentBaseResponse": {
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "integer",
            "format": "int32"
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "enrolledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Base enrollment response — no navigation properties."
      },
      "EnrollmentBaseResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/EnrollmentBaseResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "ObjectPagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Pagination metadata wrapper."
      },
      "ObjectPagedResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ObjectPagedResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
    }
  }
}